requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4

# LangChain packages
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-google-genai==0.0.6

# Vector store - Using newer version compatible with onnxruntime
chromadb==0.4.24

# PDF processing - Complete suite
unstructured==0.11.8
unstructured-inference==0.7.23
pypdf==3.17.4
pdfminer.six==20221105
pdf2image==1.16.3
pytesseract==0.3.10

# Image processing
pillow==10.1.0
opencv-python==4.8.1.78

# ML and visualization for unstructured-inference
matplotlib==3.8.2
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.3.2
# onnxruntime - let it auto-resolve to satisfy both chromadb (>=1.14.1) and unstructured-inference (<1.16)
onnxruntime==1.15.1
layoutparser==0.3.4

# Deep learning (required by unstructured-inference)
torch==2.1.2
torchvision==0.16.2

# OCR and document processing
lxml==5.1.0
beautifulsoup4==4.12.3

# Additional utilities
transformers==4.36.2
huggingface-hub==0.20.2